Se cambia de componentes la tarjeta
